// frontend/src/app/api/customGraphql.service.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: any }> = { [K in keyof T]: T[K] };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    /** The `Upload` scalar type represents a file upload. */
    Upload: any;
};
// ...
export type StockDetails = {
    __typename?: 'StockDetails';
    id: Scalars['String'];
    basicInfo?: Maybe<BasicInfo>;
    chartInfo?: Maybe<ChartInfo>;
    dividend?: Maybe<Dividend>;
    analysis?: Maybe<Analysis>;
    balanceSheet?: Maybe<BalanceSheet>;
    cashFlow?: Maybe<CashFlow>;
    overview?: Maybe<OverView>;
    perShare?: Maybe<PerShare>;
    valuation?: Maybe<Valuation>;
    financialStrength?: Maybe<FinancialStrength>;
    financialReports?: Maybe<Array<Maybe<FinancialReportNames>>>;
    financialStrengthRatio?: Maybe<FinancialStrengthRatio>;
    incomeStatement?: Maybe<IncomeStatement>;
    recommendation?: Maybe<Array<Maybe<Recommendations>>>;
    stockNewsSnippets?: Maybe<Array<Maybe<StockArticle>>>;
};

export type PerShare = {
    __typename?: 'PerShare';
    bookValuePerShareMRQ?: Maybe<Scalars['Float']>;
    cashFlowPerShareAnnual?: Maybe<Scalars['Float']>;
    cashFlowPerShareTTM?: Maybe<Scalars['Float']>;
    cashPerSharePerShareAnnual?: Maybe<Scalars['Float']>;
    cashPerSharePerShareQuarterly?: Maybe<Scalars['Float']>;
    ebitdPerShareTTM?: Maybe<Scalars['Float']>;
    freeCashFlowPerShareTTM?: Maybe<Scalars['Float']>;
    revenuePerShareTTM?: Maybe<Scalars['Float']>;
    tangibleBookValuePerShareQuarterly?: Maybe<Scalars['Float']>;
    totalCashPerShareMRQ?: Maybe<Scalars['Float']>;
};

export type FinancialStrengthRatio = {
    __typename?: 'FinancialStrengthRatio';
    currentRatioQuarterly?: Maybe<Scalars['Float']>;
    freeOperatingCashFlowToRevenue5Y?: Maybe<Scalars['Float']>;
    longTermDebtToEquityQuarterly?: Maybe<Scalars['Float']>;
    quickRatioQuarterly?: Maybe<Scalars['Float']>;
    totalDebtToEquityAnnual?: Maybe<Scalars['Float']>;
    totalDebtToEquityQuarterly?: Maybe<Scalars['Float']>;
};

export type FinancialStrength = {
    __typename?: 'FinancialStrength';
    ebitda?: Maybe<Scalars['String']>;
    grossProfitTTM?: Maybe<Scalars['String']>;
    leveredFreeCashFlowTTM?: Maybe<Scalars['String']>;
    marketCap?: Maybe<Scalars['String']>;
    netIncomeAvitoCommonTTM?: Maybe<Scalars['String']>;
    operatingCashFlowTTM?: Maybe<Scalars['String']>;
    revenueTTM?: Maybe<Scalars['String']>;
    totalCashMRQ?: Maybe<Scalars['String']>;
    totalDebtMRQ?: Maybe<Scalars['String']>;
};

export type FinancialReportNames = {
    __typename?: 'FinancialReportNames';
    collection?: Maybe<Scalars['String']>;
    name?: Maybe<Scalars['String']>;
};

export type Recommendations = {
    __typename?: 'Recommendations';
    buy?: Maybe<Scalars['Float']>;
    hold?: Maybe<Scalars['Float']>;
    period?: Maybe<Scalars['String']>;
    sell?: Maybe<Scalars['Float']>;
    strongBuy?: Maybe<Scalars['Float']>;
    strongSell?: Maybe<Scalars['Float']>;
    symbol?: Maybe<Scalars['String']>;
};

export type IncomeStatement = {
    __typename?: 'IncomeStatement';
    Basic_Average_Shares?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Basic_EPS?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Cost_of_Revenue?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Diluted_Average_Shares?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Diluted_EPS?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Diluted_NI_Available_to_Com_Stockholders?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Gross_Profit?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Interest_Expense?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Income_Common_Stockholders?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Income_from_Continuing_And_Discontinued_Operation?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Income_from_Continuing_Operation_Net_Minority_Interest?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Interest_Income?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Non_Operating_Interest_Income_Expense?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Normalized_EBITDA?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Normalized_Income?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Operating_Expense?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Operating_Income?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Other_Income_Expense?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Pretax_Income?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Reconciled_Cost_of_Revenue?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Reconciled_Depreciation?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Tax_Effect_of_Unusual_Items?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Tax_Provision?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Tax_Rate_for_Calcs?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Expenses?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Operating_Income_as_Reported?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Revenue?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Unusual_Items?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Unusual_Items_Excluding_Goodwill?: Maybe<Array<Maybe<Scalars['Float']>>>;
    date?: Maybe<Array<Maybe<Scalars['String']>>>;
};

export type CashFlow = {
    __typename?: 'CashFlow';
    Capital_Expenditure?: Maybe<Array<Maybe<Scalars['Float']>>>;
    End_Cash_Position?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Financing_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Free_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Investing_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Issuance_of_Capital_Stock?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Issuance_of_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Operating_Cash_Flow?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Repayment_of_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Repurchase_of_Capital_Stock?: Maybe<Array<Maybe<Scalars['Float']>>>;
    date?: Maybe<Array<Maybe<Scalars['String']>>>;
};

export type BalanceSheet = {
    __typename?: 'BalanceSheet';
    Capital_Lease_Obligations?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Common_Stock_Equity?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Invested_Capital?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Net_Tangible_Assets?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Ordinary_Shares_Float?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Share_Issued?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Tangible_Book_Value?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Assets?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Capitalization?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Debt?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Equity_Gross_Minority_Interest?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Total_Liabilities_Net_Minority_Interest?: Maybe<Array<Maybe<Scalars['Float']>>>;
    Working_Capital?: Maybe<Array<Maybe<Scalars['Float']>>>;
    date?: Maybe<Array<Maybe<Scalars['String']>>>;
};
// ...
export type Analysis = {
    __typename?: 'Analysis';
    growthEstimatesPercent?: Maybe<GrowthEstimatesPercent>;
    revenueEstimate?: Maybe<RevenueEstimate>;
    earnings?: Maybe<Earnings>;
};

export type ChartInfo = {
    __typename?: 'ChartInfo';
    assetsToDebtInfo?: Maybe<AssetsToDebtInfo>;
    equityToAssets?: Maybe<EquityToAssets>;
    effectiveness?: Maybe<Effectiveness>;
    margin?: Maybe<Margin>;
    otherGrowthInformation?: Maybe<OtherGrowthInformation>;
    profitMargin?: Maybe<ProfitMargin>;
};

export type ProfitMargin = {
    __typename?: 'ProfitMargin';
    expenseTTM?: Maybe<Scalars['Float']>;
    netProfitMarginTTM?: Maybe<Scalars['Float']>;
};
// ...
export type Margin = {
    __typename?: 'Margin';
    grossMarginTTM?: Maybe<Scalars['Float']>;
    netMarginGrowth5Y?: Maybe<Scalars['Float']>;
    netProfitMargin5Y?: Maybe<Scalars['Float']>;
    netProfitMarginTTM?: Maybe<Scalars['Float']>;
    operatingMarginTTM?: Maybe<Scalars['Float']>;
    pretaxMarginTTM?: Maybe<Scalars['Float']>;
};

export type Effectiveness = {
    __typename?: 'Effectiveness';
    returnOnAssetsTTM?: Maybe<Scalars['Float']>;
    returnOnEquityTTM?: Maybe<Scalars['Float']>;
    returnOnInvestmentsTTM?: Maybe<Scalars['Float']>;
};

export type AssetsToDebtInfo = {
    __typename?: 'AssetsToDebtInfo';
    totalAssets?: Maybe<Scalars['Float']>;
    totalDebt?: Maybe<Scalars['Float']>;
    prctDiff?: Maybe<Scalars['Float']>;
};

export type EquityToAssets = {
    __typename?: 'EquityToAssets';
    totalAssets?: Maybe<Scalars['Float']>;
    totalEquity?: Maybe<Scalars['Float']>;
    prctDiff?: Maybe<Scalars['Float']>;
};

export type Valuation = {
    __typename?: 'Valuation';
    currentEvToFreeCashFlowAnnual?: Maybe<Scalars['Float']>;
    currentEvToFreeCashFlowTTM?: Maybe<Scalars['Float']>;
    enterpriseValueToEBITDA?: Maybe<Scalars['Float']>;
    enterpriseValueToRevenue?: Maybe<Scalars['Float']>;
    forwardPE?: Maybe<Scalars['Float']>;
    customPE?: Maybe<Scalars['Float']>;
    peRatioTTM?: Maybe<Scalars['Float']>;
    pegRatioFiveYearExpected?: Maybe<Scalars['Float']>;
    priceToBookMRQ?: Maybe<Scalars['Float']>;
    priceToSalesTTM?: Maybe<Scalars['Float']>;
    trailingPE?: Maybe<Scalars['Float']>;
};

export type OverView = {
    __typename?: 'OverView';
    currentPrice?: Maybe<Scalars['Float']>;
    currentPriceChange?: Maybe<Scalars['Float']>;
    previousClose?: Maybe<Scalars['Float']>;
    weekHigh52?: Maybe<Scalars['Float']>;
    weekLow52?: Maybe<Scalars['Float']>;
    symbol?: Maybe<Scalars['String']>;
    earningsDate?: Maybe<Scalars['String']>;
    exDividendDate?: Maybe<Scalars['String']>;
    forwardDividendAndYield?: Maybe<Scalars['String']>;
    targetEst1y?: Maybe<Scalars['Float']>;
    targetEst1yPercent?: Maybe<Scalars['Float']>;
};

export type Dividend = {
    __typename?: 'Dividend';
    dividendDate?: Maybe<Scalars['String']>;
    dividendPerShare5Y?: Maybe<Scalars['Float']>;
    dividendPerShareAnnual?: Maybe<Scalars['Float']>;
    exDividendDate?: Maybe<Scalars['String']>;
    fiveYearAverageDividendYield?: Maybe<Scalars['String']>;
    forwardAnnualDividendRate?: Maybe<Scalars['String']>;
    forwardAnnualDividendYield?: Maybe<Scalars['String']>;
    payoutRatio?: Maybe<Scalars['String']>;
    trailingAnnualDividendRate?: Maybe<Scalars['String']>;
    trailingAnnualDividendYield?: Maybe<Scalars['String']>;
};

export type BasicInfo = {
    __typename?: 'BasicInfo';
    address1?: Maybe<Scalars['String']>;
    city?: Maybe<Scalars['String']>;
    country?: Maybe<Scalars['String']>;
    exchangeTimezoneName?: Maybe<Scalars['String']>;
    fullTimeEmployees?: Maybe<Scalars['Float']>;
    industry?: Maybe<Scalars['String']>;
    sector?: Maybe<Scalars['String']>;
    shortName?: Maybe<Scalars['String']>;
    state?: Maybe<Scalars['String']>;
    website?: Maybe<Scalars['String']>;
    zip?: Maybe<Scalars['String']>;
    netIncomeEmployeeAnnual?: Maybe<Scalars['Float']>;
    revenueEmployeeAnnual?: Maybe<Scalars['Float']>;
    sharesOutstanding?: Maybe<Scalars['Float']>;
    summary?: Maybe<Scalars['String']>;
    symbol?: Maybe<Scalars['String']>;
    logoUrl?: Maybe<Scalars['String']>;
};

export type StockArticleWrapper = {
    __typename?: 'StockArticleWrapper';
    stockNews?: Maybe<Array<Maybe<StockArticle>>>;
};

export type StockArticle = {
    __typename?: 'StockArticle';
    datetime?: Maybe<Scalars['String']>;
    headline?: Maybe<Scalars['String']>;
    image?: Maybe<Scalars['String']>;
    related?: Maybe<Scalars['String']>;
    source?: Maybe<Scalars['String']>;
    sourceName?: Maybe<Scalars['String']>;
    summary?: Maybe<Scalars['String']>;
    url?: Maybe<Scalars['String']>;
};
// ...
export type StockMainDetailsFragment = (
    { __typename?: 'StockDetails' }
    & Pick<StockDetails, 'id'>
    & {
    overview?: Maybe<(
        { __typename?: 'OverView' }
        & Pick<OverView, 'symbol' | 'currentPrice' | 'weekHigh52' | 'weekLow52' | 'previousClose' | 'earningsDate' | 'exDividendDate' | 'forwardDividendAndYield' | 'targetEst1y'>
        )>, basicInfo?: Maybe<(
        { __typename?: 'BasicInfo' }
        & Pick<BasicInfo, 'sector' | 'industry' | 'logoUrl' | 'shortName' | 'website'>
        )>
}
    );